- fix loop issue when checking checksum

+++ infrastructure-provisioning/aws/src/emr/scripts/create_configs.py
@@ -61,10 +61,12 @@ def install_emr_spark(args):
     s3_client = boto3.client('s3', config=Config(signature_version='s3v4'), region_name=args.region)
     s3_client.download_file(args.bucket, args.user_name + '/' + args.cluster_name + '/spark.tar.gz', '/tmp/spark.tar.gz')
     s3_client.download_file(args.bucket, args.user_name + '/' + args.cluster_name + '/spark-checksum.chk', '/tmp/spark-checksum.chk')
-    while 'WARNING' in checksum_check('/tmp/spark-checksum.chk'):
+    if 'WARNING' in checksum_check('/tmp/spark-checksum.chk'):
         local('rm -f /tmp/spark.tar.gz')
         s3_client.download_file(args.bucket, args.user_name + '/' + args.cluster_name + '/spark.tar.gz', '/tmp/spark.tar.gz')
-        checksum_check('/tmp/spark-checksum.chk')
+        if 'WARNING' in checksum_check('/tmp/spark-checksum.chk'):
+            print "The checksum of spark.tar.gz is mismatched. It could be caused by aws network issue."
+            sys.exit(1)
     local('sudo tar -zhxvf /tmp/spark.tar.gz -C /opt/' + args.emr_version + '/' + args.cluster_name + '/')
 
 
@@ -81,10 +83,12 @@ def jars(args):
     s3_client = boto3.client('s3', config=Config(signature_version='s3v4'), region_name=args.region)
     s3_client.download_file(args.bucket, 'jars/' + args.emr_version + '/jars.tar.gz', '/tmp/jars.tar.gz')
     s3_client.download_file(args.bucket, 'jars/' + args.emr_version + '/jars-checksum.chk', '/tmp/jars-checksum.chk')
-    while 'WARNING' in checksum_check('/tmp/jars-checksum.chk'):
+    if 'WARNING' in checksum_check('/tmp/jars-checksum.chk'):
         local('rm -f /tmp/jars.tar.gz')
         s3_client.download_file(args.bucket, 'jars/' + args.emr_version + '/jars.tar.gz', '/tmp/jars.tar.gz')
-        checksum_check('/tmp/jars-checksum.chk')
+        if 'WARNING' in checksum_check('/tmp/jars-checksum.chk'):
+            print "The checksum of jars.tar.gz is mismatched. It could be caused by aws network issue."
+            sys.exit(1)
     local('tar -zhxvf /tmp/jars.tar.gz -C ' + emr_dir)
